feat: Update utils export method #183
